requirements.txt
openai~=0.27.4
flask~=2.2.3
python-dotenv
langchain==0.0.137
SpeechRecognition
flask-cors
flask-serialize
pip~=23.1.2
toml~=0.10.2
wheel~=0.37.1
cryptography~=40.0.1
lxml~=4.9.2
tornado~=6.2
future~=0.18.3
requests~=2.28.2
rich~=13.0.1
colorama~=0.4.6
ipywidgets~=8.0.6
ipython~=8.12.0
notebook~=6.5.3
PyYAML~=6.0
idna~=3.4
multidict~=6.0.4
attrs~=22.2.0
MarkupSafe~=2.1.2
Werkzeug~=2.2.3
click~=8.1.3
Jinja2~=3.1.2
itsdangerous~=2.1.2
numpy~=1.23.5
setuptools~=58.1.0
pytz~=2023.3
cffi~=1.15.1
numba~=0.57.0
pydantic~=1.10.7
aiohttp~=3.8.4
pandas~=1.5.3
matplotlib~=3.7.1
scipy~=1.10.1
scikit-learn~=1.2.2
tenacity~=8.2.2
yarl~=1.8.2
aiosignal~=1.3.1
frozenlist~=1.3.3
python-multipart~=0.0.6
certifi~=2022.12.7
greenlet~=2.0.2
psutil~=5.9.4
decorator~=5.1.1
urllib3~=1.26.15
SQLAlchemy~=1.4.47
six~=1.16.0
marshmallow~=3.19.0
packaging~=23.0
soundfile~=0.12.1
pinecone-client==2.2.1
unstructured
tabulate
pdf2image
tiktoken
ndg-httpsclient
pyopenssl
pyasn1
pinecone-client
torchvision
unstructured_inference
google-cloud-storage